Moved User into dto directory
User is an object class, should be located in the dto directory
